chore(PERF): Fire analytics events on idle time if idleCallback feature is available
We have been digging into the frontend performance space for Jira and found that the browser metrics analytics events generate a high cost for users due to how the are being manage.
We've seen up to 200ms improvement to Time To Interactive times in synthetic tenants for Jira, which are a good representation of a P50 user.
During an analysis of performance support case we found that the number of events fired by connect addons outside the control of the product is considerable, leading us to propose this solution. We have already implemented this same deferral inside the product itself for certain critical path events. However, given these are connect addon events and of lower priority in the critical path we understand we can bluntly defer them all.